% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateP2G.R
\name{calculateP2G}
\alias{calculateP2G}
\title{Establish peak to gene links based on correlations between ATAC-seq peaks and RNA-seq genes}
\usage{
calculateP2G(
  ArchR_path = NULL,
  useDim = "IterativeLSI",
  useMatrix = "GeneIntegrationMatrix",
  cor_cutoff = 0.5,
  ...
)
}
\arguments{
\item{ArchR_path}{String specifying the path to an ArchR project if ArchR's implementation of addPeak2GeneLinks is desired}

\item{useDim}{String specifying the dimensional reduction representation in the ArchR project to use or the name of the reduced dimension matrix supplied by the user}

\item{useMatrix}{String specifying which the name of the gene expression matrix in the ArchR project to use.
It is often called the "GeneExpressionMatrix" for multiome and "GeneIntegrationMatrix" for unpaired data in ArchR project.}

\item{cor_cutoff}{A numeric scalar to specify the correlation cutoff between ATAC-seq peaks and RNA-seq genes to assign peak to gene links.
Default correlation cutoff is 0.5.}

\item{...}{other parameters to pass to addPeak2GeneLinks from ArchR package}
}
\value{
A DataFrame of Peak to Gene correlation
}
\description{
Establish peak to gene links based on correlations between ATAC-seq peaks and RNA-seq genes
}
\author{
Xiaosai Yao, Shang-yang Chen
}
